fix(production): align services with diagram - fix type mismatches and repo queries
- ProductService: changed addStock/reduceStock params to Double (match Product.stock)
- ProductionOrderService: fixed raw material consumption (cast Double->Integer for RawMaterial.stock)
- ProductionOrderService: updated date range queries (replace plannedDate with startDate)
- ProductionOrderRepository: replaced deprecated queries (plannedDate->startDate, old enum EN_COURS->EN_PRODUCTION)
- ProductController: updated stock endpoints to accept Double params
- BillOfMaterialMapper: fixed calculateLineCost()->calculateTotalCost() mapping
- ProductionOrder entity: added isDelayed() helper method for mapper

Build now succeeds (mvn package)

+++ supplychainx-production/src/main/java/com/supplychainx/production/controller/ProductController.java
@@ -119,7 +119,7 @@ public class ProductController {
     @Operation(summary = "Ajouter du stock", description = "Ajoute une quantité au stock d'un produit")
     public ResponseEntity<ProductResponseDTO> addStock(
             @PathVariable Long id,
-            @RequestParam Integer quantity) {
+            @RequestParam Double quantity) {
         log.info("Requête d'ajout de stock pour le produit ID: {} - Quantité: {}", id, quantity);
         ProductResponseDTO response = productService.addStock(id, quantity);
         return ResponseEntity.ok(response);
@@ -129,7 +129,7 @@ public class ProductController {
     @Operation(summary = "Réduire le stock", description = "Réduit le stock d'un produit (avec validation de disponibilité)")
     public ResponseEntity<ProductResponseDTO> reduceStock(
             @PathVariable Long id,
-            @RequestParam Integer quantity) {
+            @RequestParam Double quantity) {
         log.info("Requête de réduction de stock pour le produit ID: {} - Quantité: {}", id, quantity);
         ProductResponseDTO response = productService.reduceStock(id, quantity);
         return ResponseEntity.ok(response);
